+++ src/components/forms/AutoSwitchArrayEntry.tsx
