+++ src/components/forms/FormSection.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import styled from 'styled-components'
 import { Card, CardContent, CardHeader, Tooltip } from '@material-ui/core'
 import { Warning as WarningIcon } from '@material-ui/icons'
-import { errorSummary, useErrorsInChildren } from './utils'
 import { AutoPrefix } from './PrefixContext'
+import { errorSummary, useErrorsInChildren } from './utils'
 
 // todo: should prevent the `hasErrors` prop from being passed down
 const FormCard = styled(Card)<{ hasErrors?: boolean }>`
